pyproject.toml
[tool.poetry]
name = "bm-ce-migration"
version = "0.1.0"
description = ""
authors = ["FruitPassion <gamehotdl@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
colorama = "^0.4.6"
tqdm = "^4.66.4"
paramiko = "^3.4.0"
simple-term-menu = "^1.6.4"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
black = "^24.3.0"
isort = "^5.13.2"
build = "^1.0.3"
twine = "^4.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.black]
line-length = 200